refactor(images): remove sync_upstream_images script and integrate image syncing into generate.js
