+++ package.json
@@ -11,13 +11,11 @@
     "swizzle": "docusaurus swizzle",
     "clear": "docusaurus clear",
     "sync-upstream-scripts": "node src/sync_upstream_scripts.js",
-    "sync-upstream-images": "node src/sync_upstream_images.js",
-    "sync-upstream-images-watch": "node src/sync_upstream_images.js --watch",
-    "roll": "rimraf -g nodejs/docs/* && rimraf -g python/docs/* && rimraf -g java/docs/* && rimraf -g dotnet/docs/* && npm run sync-upstream-scripts && npm run sync-upstream-images && node src/generate.js",
-    "start-dotnet": "concurrently \"node src/generate.js --watch csharp\" \"npm run sync-upstream-images-watch\" \"npm run start\"",
-    "start-java": "concurrently \"node src/generate.js --watch java\" \"npm run sync-upstream-images-watch\" \"npm run start\"",
-    "start-nodejs": "concurrently \"node src/generate.js --watch js\" \"npm run sync-upstream-images-watch\" \"npm run start\"",
-    "start-python": "concurrently \"node src/generate.js --watch python\" \"npm run sync-upstream-images-watch\" \"npm run start\"",
+    "roll": "rimraf -g nodejs/docs/* && rimraf -g python/docs/* && rimraf -g java/docs/* && rimraf -g dotnet/docs/* && npm run sync-upstream-scripts && node src/generate.js",
+    "start-dotnet": "concurrently \"node src/generate.js --watch csharp\" \"npm run start\"",
+    "start-java": "concurrently \"node src/generate.js --watch java\" \"npm run start\"",
+    "start-nodejs": "concurrently \"node src/generate.js --watch js\" \"npm run start\"",
+    "start-python": "concurrently \"node src/generate.js --watch python\" \"npm run start\"",
     "prepare-dotnet": "npm run cleanup && cp -r dotnet/* .",
     "prepare-java": "npm run cleanup && cp -r java/* .",
     "prepare-nodejs": "npm run cleanup && cp -r nodejs/* .",
